requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.13
django-apscheduler==0.6.2
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
dj-database-url==0.5.0
django-cloudinary-storage==0.3.0
django-extra-fields==3.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.7